+++ CHANGELOG.md
@@ -227,3 +227,4 @@
 - fix: fix the event delivered to handlers in FIFO order
 - feat: add `set_config` to `EngineBuilder` and unit test
 - fix: change task id format with `{pid}{tid}`, remove `SEP` between `pid` and `tid`
+- fix: add QuickJS interrupt handle to deal with deadline issue